refactor(InputView.java): 입력 데이터 검증 로직 분리
입력한 데이터를 검증하는 로직을 InputView에서 분리

BREAKING CHANGE: InputView의 검증 로직을 InputValidator로 분리
as-is:
public class InputView {

    private static final Scanner scanner = new Scanner(System.in);
    private static final String NUMBER_PATTERN = "^[0-9]*$";
    private static final Pattern COMPILED_NUMBER_PATTERN = Pattern.compile(NUMBER_PATTERN);
    private static final String WINNING_NUMBER_PATTERN = "^([1-9]{1,2}[,][\\s]?){5}[1-9]{1,2}$";
    private static final Pattern COMPILED_WINNING_NUMBER_PATTERN = Pattern.compile(WINNING_NUMBER_PATTERN);
    private static final String WINNING_NUMBER_DISTRIBUTOR = ",";
    private static final int AMOUNT_MIN_RANGE = 1_000;
    private static final int AMOUNT_MAX_RANGE = 100_000;

    public static int inputValidMoney() {
        final String money = inputMoney();
        return Integer.parseInt(money);
    }

    private static String inputMoney() {
        try {
            ...
            validateMoney(money);
            ...
        } catch (IllegalArgumentException exception) {
            ...
        }
    }

    private static void validateMoney(final String money) {
            validateNumber(money);
            validateRange(money);
    }

    private static void validateRange(String money) {
            int number = Integer.parseInt(money);
            if (number < AMOUNT_MIN_RANGE || number > AMOUNT_MAX_RANGE) {
                    throw new IllegalArgumentException(LotteryMessage.PURCHASE_AMOUNT_RANGE_ERROR);
            }
    }

    private static void validateNumber(String money) {
            if (!COMPILED_NUMBER_PATTERN.matcher(money).matches()) {
                    throw new IllegalArgumentException(LotteryMessage.PURCHASE_AMOUNT_NOT_NUMBER_ERROR);
            }
    }

    ...

    private static String inputLotteryNumber() {
        try {
            ...
            validateWinningNumber(winningNumber);
            ...
        } catch (IllegalArgumentException exception) {
            ...
        }
    }

    private static void validateWinningNumber(String winningNumber) {
        if (!COMPILED_WINNING_NUMBER_PATTERN.matcher(winningNumber).matches()) {
            throw new IllegalArgumentException(LotteryMessage.WINNING_NUMBER_FORMAT_ERROR);
        }
    }

    ...

    private static String inputBonusNumber() {
        try {
            ...
            validateNumber(bonusNumber);
            ...
        } catch (IllegalArgumentException exception) {
            ...
        }
    }
}

to-be:
public class InputView {

    private static final Scanner scanner = new Scanner(System.in);
    private static final String WINNING_NUMBER_DISTRIBUTOR = ",";

   ...

    private static String inputMoney() {
        try {
            ...
            InputValidator.validateMoney(money);
            ...
        } catch (IllegalArgumentException exception) {
            ...
        }
    }

   ...

    private static String inputLotteryNumber() {
        try {
            ...
            InputValidator.validateWinningNumber(winningNumber);
            ...
        } catch (IllegalArgumentException exception) {
            ...
        }
    }

   ...

    private static String inputBonusNumber() {
        try {
            ...
            InputValidator.validateNumber(bonusNumber);
            ...
        } catch (IllegalArgumentException exception) {
            ...
        }
    }
}
